adds warning message if user is not logged in

+++ src/components/VotesUpOrDown.js
@@ -1,10 +1,80 @@
 import React, { useEffect, useState } from "react";
+import { useTravelStore } from "../TipsContext";
+import { observer } from "mobx-react";
+import { useHistory } from "react-router-dom";
+import ReactDOM from "react-dom";
+
+const useModal = () => {
+  const [isShowing, setIsShowing] = useState(false);
+  function toggle() {
+    setIsShowing(!isShowing);
+  }
+
+  return {
+    isShowing,
+    toggle,
+  };
+};
+
+const Modal = ({ isShowing, hide, user, handlesLogin, handlesRegister }) =>
+  isShowing
+    ? ReactDOM.createPortal(
+        <React.Fragment>
+          <div className="modal-overlay" />
+          <div
+            className="modal-wrapper"
+            aria-modal
+            aria-hidden
+            tabIndex={-1}
+            role="dialog"
+          >
+            <div className="modal">
+              <div className="modal-header">
+                <button
+                  type="button"
+                  id="x-btn"
+                  className="modal-close-button"
+                  data-dismiss="modal"
+                  aria-label="Close"
+                  onClick={hide}
+                >
+                  <i className="fas fa-window-close"></i>
+                </button>
+              </div>
+              <div className="inner-modal">
+                <i className="fas fa-exclamation-triangle fa-3x"></i>
+              </div>
+              <h4>You must be logged in to vote.</h4>
+
+              <div className="delete-account-btns">
+                <button type="button" id="no-btn" onClick={handlesLogin}>
+                  Login
+                </button>
+                <button id="yes-btn" onClick={handlesRegister}>
+                  Register
+                </button>
+              </div>
+            </div>
+          </div>
+        </React.Fragment>,
+        document.body
+      )
+    : null;
 
 function VotesUpOrDown(props) {
   const [direction, setDirection] = useState(0);
   const [votes, setVotes] = useState(0);
-
+  const store = useTravelStore();
+  let history = useHistory();
+  const { isShowing, toggle } = useModal();
   // needs to warn user they must to sign up to vote
+
+  function handlesLogin() {
+    history.push("/login");
+  }
+  function handlesRegister() {
+    history.push("/register");
+  }
 
   useEffect(() => {
     async function fetchDirection() {
@@ -23,6 +93,10 @@ function VotesUpOrDown(props) {
 
   let handlesVote = (desiredDirection) => {
     return async () => {
+      if (!store.user) {
+        toggle();
+        return;
+      }
       if (desiredDirection === direction) {
         await fetch(`/suggestions/${props.suggestion?.id}/vote`, {
           method: "DELETE",
@@ -57,10 +131,17 @@ function VotesUpOrDown(props) {
       setVotes(data);
     };
     getAmountOfVotes();
-  }, [direction, props.suggestion?.id]);
+  }, [direction, props.suggestion?.id, setVotes]);
 
   return (
     <div className="votes-container">
+      <Modal
+        isShowing={isShowing}
+        hide={toggle}
+        user={store?.user}
+        handlesLogin={handlesLogin}
+        handlesRegister={handlesRegister}
+      />
       <button
         className={`button-up ${direction === 1 ? "active" : ""}`}
         onClick={doUpvote}
@@ -78,4 +159,4 @@ function VotesUpOrDown(props) {
   );
 }
 
-export default VotesUpOrDown;
+export default observer(VotesUpOrDown);
